Add Kebab, move dropdowns and exceptions in separate modules
